change all int for element id to Long

+++ src/main/java/com/example/Create_Character/repos/ElementRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ElementRepo extends JpaRepository<Element, Integer> {
+public interface ElementRepo extends JpaRepository<Element, Long> {
 }
